requirements.txt
google-auth-oauthlib==1.1.0
google-api-core==2.14.0
google-api-python-client==2.107.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.61.0
